+++ talpid-core/src/tunnel_state_machine/connecting_state.rs
@@ -194,6 +194,13 @@ impl ConnectingState {
         if let Err(error) = shared_values.route_manager.clear_routes() {
             log::error!("{}", error.display_chain_with_msg("Failed to clear routes"));
         }
+        #[cfg(target_os = "linux")]
+        if let Err(error) = shared_values.route_manager.clear_routing_rules() {
+            log::error!(
+                "{}",
+                error.display_chain_with_msg("Failed to clear routing rules")
+            );
+        }
     }
 
     fn disconnect(
@@ -354,14 +361,6 @@ impl TunnelState for ConnectingState {
                         ErrorStateCause::SetFirewallPolicyError(error),
                     )
                 } else {
-                    #[cfg(target_os = "linux")]
-                    if let Err(error) = shared_values.route_manager.enable_exclusions_routes() {
-                        error!(
-                            "{}",
-                            error.display_chain_with_msg("Failed to set up split tunneling")
-                        );
-                    }
-
                     #[cfg(target_os = "android")]
                     {
                         if retry_attempt > 0 && retry_attempt % MAX_ATTEMPTS_WITH_SAME_TUN == 0 {